chore: refactor test tests into ecosystems
