+++ package.json
@@ -27,6 +27,7 @@
     "find-circular": "npm run build && madge --circular ./dist",
     "format": "prettier --write '{src,test,scripts}/**/*.{js,ts}'",
     "prepare": "npm run build",
+    "test:acceptance": "tap test/acceptance/cli-test.nuget.test.ts -Rspec --timeout=300 --node-arg=-r --node-arg=ts-node/register",
     "tap": "tap test/*.test.* test/acceptance/*.test.* test/system/*.test.* -Rspec --timeout=300 --node-arg=-r --node-arg=ts-node/register",
     "test": "npm run test-common && npm run tap",
     "test-common": "npm run check-tests && npm run build && npm run lint && node --require ts-node/register src/cli test --org=snyk",
