+++ test/acceptance/cli-protect.test.ts
